Refactor summaries code in package to avoid circular deps
